+++ src/lib/config.ts
@@ -85,7 +85,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "4.0.0",
-  sdkVersion: "5.2.0-alpha.65",
-  genVersion: "2.690.1",
-  userAgent: "speakeasy-sdk/typescript 5.2.0-alpha.65 2.690.1 4.0.0 meep-moop",
+  sdkVersion: "5.2.0-alpha.66",
+  genVersion: "2.698.4",
+  userAgent: "speakeasy-sdk/typescript 5.2.0-alpha.66 2.698.4 4.0.0 meep-moop",
 } as const;