+++ RELEASES.md
@@ -783,3 +783,13 @@ Based on:
 - [typescript v5.2.0-alpha.62] .
 ### Releases
 - [NPM v5.2.0-alpha.62] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.62 - .
+
+## 2025-09-11 12:09:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.615.2 (2.698.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v5.2.0-alpha.66] .
+### Releases
+- [NPM v5.2.0-alpha.66] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.66 - .